removed last vestiges of static link classes
